Fixed project structure. Moved order aggregate classes from aggregate.aggregate package to aggregate.
