+++ api-java/src/main/java/javaclasses/mealorder/c/context/BoundedContexts.java
@@ -24,7 +24,7 @@ import com.google.common.annotations.VisibleForTesting;
 import io.spine.server.BoundedContext;
 import io.spine.server.storage.StorageFactory;
 import io.spine.server.storage.memory.InMemoryStorageFactory;
-import javaclasses.mealorder.c.aggregate.repository.OrderRepository;
+import javaclasses.mealorder.c.repository.OrderRepository;
 import javaclasses.mealorder.c.repository.VendorRepository;
 
 import static com.google.common.base.Preconditions.checkNotNull;
